see #904, added an option not to remove pagefile from the backup

+++ Migrate/Windows/WindowsBackupAdjust.py
@@ -579,8 +579,9 @@ class WindowsBackupAdjust(BackupAdjust.BackupAdjust):
 
         # 1) Exclude files not needed
         # TODO: find all the pagefiles
-        self.removeFile("pagefile.sys")
-        self.removeFile("hiberfil.sys")
+        if self.__adjustConfig.removePageFile():
+            self.removeFile("pagefile.sys")
+            self.removeFile("hiberfil.sys")
 
         # 1) Replace files we gonna change
         #TODO: add auto-add data to an any extent we read
